Split pattern/icu
